requirements.txt
flask>=2.3.0
flask-socketio>=5.3.0
python-socketio>=5.9.0
python-engineio>=4.8.0
eventlet>=0.35.0
dataclasses-json>=0.6.0
python-dotenv>=1.0.0
PyQt6>=6.5.0
PyQt6-WebEngine>=6.5.0
psutil>=5.9.5
requests>=2.31.0
pillow>=10.0.0
keyboard>=0.13.5
websocket-client>=1.6.1
cx_Freeze>=6.15.0
pywin32>=306